Require C extensions for successful build
This behavior is being changed to improve the user experience when installing giraffez from pip.  Pip provides no way to output information that would indicate which of the C extensions did or did not install correctly so for now it must simply fail or succeed when installing (which produces retroactively the setuptools output).