Improve build tools and graceful shutdown
This adds a new API function `register_graceful_shutdown_signal` and only uses in `commandline.py`. This ensures that those that import the library will not have that C signal registered unless they wish to register it with the newly exposed function. As part of making this improvement the build system has been changed yet again to allow for more flexibility in how the various C extension modules in giraffez are defined and built.