Add methods to Row/Result to replace to_json
These methods are functionally the same but the new names are just more indicative of what is actually happening here.  The old to_json has been kept and changed to point to the new methods to not break the API and a deprecation label has been added to the docs since the new methods should be used going forward.