Fix multiple Cmd bugs, disable parallel build
In `giraffez/cli/cmdobject.c` the session was getting renewed over and over again causing it to be much slower than normal and also to eventually fail with something like 'CLIv2 init request failed'.  This was remedied by ensuring that fetch_record did not set the dbcarea in a way that sets up a new session internally with CLIv2.  The other issue was running multiple Cmd instances in a single script.  This bug was introduced when calling `Cmd_close` also within the dealloc function.  This was wrong because it ran the CLIv2 function for cleanup which apparently uses global variables to store some data (like buffers or something like that).

The other issue was with building on >= Python 3.5 getting strange errors about "truncated files" or just hanging forever on Windows.  The issue was identified to be the parallel option being set when compiling.