[WIP] Overhaul of Cmd/Export, C styling changes
There is a better abstraction for both Cmd and Export.  Cmd is going to be more usable by other C programs and also uses the prepare statement mode to improve the speed and cost of calls to `Cmd.get_columns`.  Export greatly simplified, columns are now handled in C and not passed back and forth from C to the Python interpreter.  The C style has changed to put all the C source files in `giraffez/src` and switch the pointer side to the right (previously the left).