Cleanup C/C++ extension, add row count, possible fix for macOS segfault
On macOS only, using BulkLoad causes the program to exit with
`Segmentation Fault: 11`. This segfault is related to differences in
either the platform or the code that is produced causing the Connection
object to not be available. The base class has been changed to be
composed in the parent class to have better control over the lifetime
while figuring out specifically what is causing the issue.